+++ nodemapper/src/Terminal/TerminalController.ts
@@ -22,7 +22,7 @@ class TerminalController {
   constructor() {
     this.term = new Terminal();
     this.fitAddon = new FitAddon();
-    this.term.loadAddon(this.fitAddon);
+    //this.term.loadAddon(this.fitAddon);
     if (window !== undefined) this.terminalAPI = window.terminalAPI;
   }
 
@@ -42,7 +42,8 @@ class TerminalController {
       // Connect terminal to the (electron) API (but only if it exists)
       this.term.open(this.xtermRef.current);
       this.term.resize(80, 5);
-      this.fitAddon.fit();  // TODO: Not resizing terminal as intended
+      //this.fitAddon.fit();  // TODO: Not resizing terminal as intended
+                              // Infact, this breaks the terminal
       this.term.onData((data) => this.terminalAPI.sendData(data));
       this.terminalAPI.receiveData((event, data) => this.sendData(data));
     }
