+++ nodemapper/src/redux/reducers/builder.ts
@@ -15,6 +15,7 @@ interface IBuilderState {
   conda_backend: string;
   environment_variables: string;
   auto_validate_connections: boolean;
+  config_pane_open: boolean;
 }
 
 // State
@@ -39,6 +40,7 @@ const builderStateInit: IBuilderState = {
   conda_backend: "builtin", // builtin | system
   environment_variables: "",
   auto_validate_connections: false,
+  config_pane_open: false,
 };
 
 // Nodemap
@@ -61,10 +63,12 @@ const builderReducer = createReducer(builderStateInit, (builder) => {
     })
     .addCase(actions.builderNodeSelected, (state, action) => {
       // Action intercepted in middleware to control display
+      state.config_pane_open = true;
       console.info("[Reducer] " + action.type);
     })
     .addCase(actions.builderNodeDeselected, (state, action) => {
       // Action intercepted in middleware to control display
+      state.config_pane_open = state.settings_visible;
       console.info("[Reducer] " + action.type);
     })
     .addCase(actions.builderGetRemoteModules, (state, action) => {
@@ -101,10 +105,12 @@ const builderReducer = createReducer(builderStateInit, (builder) => {
     })
     .addCase(actions.builderSetSettingsVisibility, (state, action) => {
       state.settings_visible = action.payload;
+      state.config_pane_open = state.settings_visible;
       console.info("[Reducer] " + action.type);
     })
     .addCase(actions.builderToggleSettingsVisibility, (state, action) => {
       state.settings_visible = !state.settings_visible;
+      state.config_pane_open = state.settings_visible;
       console.info("[Reducer] " + action.type);
     })
     .addCase(actions.builderSetSnakemakeArgs, (state, action) => {
